server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["John MacKenzie <john.mackenzie@icf.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.12.1"
requests = "^2.24.0"
databases = {extras = ["postgresql"], version = "^0.3.2"}
sqlalchemy = "^1.3.20"
alembic = "^1.4.3"
python-dotenv = "^0.14.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
email-validator = "^1.1.1"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]
mypy = "^0.790"
flake8 = "^3.8.4"
autoflake = "^1.4"
isort = "^5.6.4"
black = "^20.8b1"
pytest = "^6.1.1"
pytest-cov = "^2.10.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

